Quick-n-dirty podman integration
Lies abound in this commit. I'm not getting the list of networks, I'm
getting a list of images, and I'm not even returning that, I'm just
printing it out. I hacked this together from a few blog posts and Chat
GPT 4o. It's awful in many directions.

That said, I'm new to go, and I have a think talking to a thing, so
that's progress.